added Wiki based content support and basic refactoring (#47)
* added support for Wiki based content
