+++ backend/Generator/__init__.py
@@ -1,3 +1,3 @@
 # Constructor for questgen
 from __future__ import absolute_import
-from Generator.main import MCQGenerator, BoolQGenerator, ShortQGenerator, AnswerPredictor, GoogleDocsService, FileProcessor
+from Generator.main import MCQGenerator, BoolQGenerator, ShortQGenerator, AnswerPredictor, GoogleDocsService, FileProcessor, QuestionGenerator
