+++ backend/server.py
@@ -17,9 +17,8 @@ import webbrowser
 from apiclient import discovery
 from httplib2 import Http
 from oauth2client import client, file, tools
-from Generator.question_generator import QuestionGenerator
+from mediawikiapi import MediaWikiAPI
 
-qg = QuestionGenerator()
 app = Flask(__name__)
 CORS(app)
 print("Starting Flask App...")
@@ -31,16 +30,26 @@ MCQGen = main.MCQGenerator()
 answer = main.AnswerPredictor()
 BoolQGen = main.BoolQGenerator()
 ShortQGen = main.ShortQGenerator()
+qg = main.QuestionGenerator()
 docs_service = main.GoogleDocsService(SERVICE_ACCOUNT_FILE, SCOPES)
 file_processor = main.FileProcessor()
+mediawikiapi = MediaWikiAPI()
 qa_model = pipeline("question-answering")
+
+
+def process_input_text(input_text, use_mediawiki):
+    if use_mediawiki == 1:
+        input_text = mediawikiapi.summary(input_text,8)
+    return input_text
 
 
 @app.route("/get_mcq", methods=["POST"])
 def get_mcq():
     data = request.get_json()
     input_text = data.get("input_text", "")
+    use_mediawiki = data.get("use_mediawiki", 0)
     max_questions = data.get("max_questions", 4)
+    input_text = process_input_text(input_text, use_mediawiki)
     output = MCQGen.generate_mcq(
         {"input_text": input_text, "max_questions": max_questions}
     )
@@ -52,7 +61,9 @@ def get_mcq():
 def get_boolq():
     data = request.get_json()
     input_text = data.get("input_text", "")
+    use_mediawiki = data.get("use_mediawiki", 0)
     max_questions = data.get("max_questions", 4)
+    input_text = process_input_text(input_text, use_mediawiki)
     output = BoolQGen.generate_boolq(
         {"input_text": input_text, "max_questions": max_questions}
     )
@@ -64,7 +75,9 @@ def get_boolq():
 def get_shortq():
     data = request.get_json()
     input_text = data.get("input_text", "")
+    use_mediawiki = data.get("use_mediawiki", 0)
     max_questions = data.get("max_questions", 4)
+    input_text = process_input_text(input_text, use_mediawiki)
     output = ShortQGen.generate_shortq(
         {"input_text": input_text, "max_questions": max_questions}
     )
@@ -76,9 +89,11 @@ def get_shortq():
 def get_problems():
     data = request.get_json()
     input_text = data.get("input_text", "")
+    use_mediawiki = data.get("use_mediawiki", 0)
     max_questions_mcq = data.get("max_questions_mcq", 4)
     max_questions_boolq = data.get("max_questions_boolq", 4)
     max_questions_shortq = data.get("max_questions_shortq", 4)
+    input_text = process_input_text(input_text, use_mediawiki)
     output1 = MCQGen.generate_mcq(
         {"input_text": input_text, "max_questions": max_questions_mcq}
     )
@@ -305,6 +320,8 @@ def generate_gform():
 def get_shortq_hard():
     data = request.get_json()
     input_text = data.get("input_text", "")
+    use_mediawiki = data.get("use_mediawiki", 0)
+    input_text = process_input_text(input_text,use_mediawiki)
     input_questions = data.get("input_question", [])
     output = qg.generate(
         article=input_text, num_questions=input_questions, answer_style="sentences"
@@ -316,6 +333,8 @@ def get_shortq_hard():
 def get_mcq_hard():
     data = request.get_json()
     input_text = data.get("input_text", "")
+    use_mediawiki = data.get("use_mediawiki", 0)
+    input_text = process_input_text(input_text,use_mediawiki)
     input_questions = data.get("input_question", [])
     output = qg.generate(
         article=input_text, num_questions=input_questions, answer_style="multiple_choice"
