+++ extension/package.json
@@ -13,7 +13,9 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.2.1",
-    "react-router-dom": "^6.23.1"
+    "react-router-dom": "^6.23.1",
+    "react-switch": "^7.0.0",
+    "react-toggle-switch": "^3.0.4"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
